fix : #3 조회 API 위치 변경
