+++ apiTest/Liked.http
@@ -1,15 +1,15 @@
-### 관심 매장 등록 및 취소
+### 매장 좋아요 등록 및 취소
 POST http://localhost:8080/api/v1/users/liked/stores/1
 Authorization: {{user1Auth}}
 
-### 관심 매장 등록 및 취소 (존재하지 않는 매장)
+### 매장 좋아요 등록 및 취소 (존재하지 않는 매장)
 POST http://localhost:8080/api/v1/users/liked/stores/100
 Authorization: {{user1Auth}}
 
 
 
 ### 리뷰 좋아요 등록 및 취소
-POST http://localhost:8080/api/v1/users/liked/reviews/1
+POST http://localhost:8080/api/v1/users/liked/reviews/3
 Authorization: {{user2Auth}}
 
 ### 리뷰 좋아요 등록 및 취소 (존재하지 않는 리뷰)
@@ -22,6 +22,10 @@ Authorization: {{user1Auth}}
 
 
 
-### 관심 매장 전체 조회
+### 좋아요한 매장 전체 조회
 GET http://localhost:8080/api/v1/users/liked/stores
 Authorization: {{user1Auth}}
+
+### 좋아요한 유저 리뷰 전체 조회
+GET http://localhost:8080/api/v1/users/liked/userReviews
+Authorization: {{user2Auth}}
