+++ apiTest/UserReviews.http
@@ -1,3 +1,3 @@
-### 좋아요 누른 리뷰 전체 조회하기
-GET http://localhost:8080/api/v1/reviews/consumers/liked
-Authorization: {{user2Auth}}
+#### 좋아요 누른 리뷰 전체 조회하기
+#GET http://localhost:8080/api/v1/reviews/consumers/liked
+#Authorization: {{user2Auth}}
